Rename classes and testing files
